Move copy yesterday button to tasklist and show only when empty
