+++ src/components/TaskList/index.js
@@ -3,14 +3,25 @@ import PropTypes from 'prop-types';
 import TaskListItemContainer from '../TaskListItem/container';
 import './style.css';
 
-const TaskList = ({ tasks }) => {
-  const createTaskListItem = task => <TaskListItemContainer key={task.id} task={task} />;
-  const activeTasks = tasks.filter(task => !task.completed).map(createTaskListItem);
-  const completedTasks = tasks.filter(task => task.completed).map(createTaskListItem);
+const TaskList = ({ tasks, copyYesterday }) => {
+  const sortedTasks = tasks.map(task => <TaskListItemContainer key={task.id} task={task} />)
+    .sort((a, b) => {
+      const aCompleted = a.props.task.completed;
+      const bCompleted = b.props.task.completed;
+      if (aCompleted !== bCompleted) {
+        return aCompleted - bCompleted;
+      }
+      return a.key - b.key;
+    });
+
   return (
     <div className="task-list">
-      {activeTasks}
-      {completedTasks}
+      {sortedTasks}
+      {sortedTasks.length === 0 && (
+        <button type="button" onClick={copyYesterday}>
+          Copy from Yesterday
+        </button>
+      )}
     </div>
   );
 };
@@ -19,6 +30,7 @@ TaskList.propTypes = {
   tasks: PropTypes.arrayOf(PropTypes.shape({
     id: PropTypes.number.isRequired,
   })).isRequired,
+  copyYesterday: PropTypes.func.isRequired,
 };
 
 export default TaskList;
